Refactor: move MuzzleFlash to `src` module; add type hints
